Endpoint Funcionario

+++ src/main/java/com/wipro/petshop/service/FuncionarioService.java
@@ -1,16 +1,20 @@
 package com.wipro.petshop.service;
 
 import com.wipro.petshop.entity.Funcionario;
+import com.wipro.petshop.exception.NotFoundException;
 import com.wipro.petshop.repository.FuncionarioRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
 
 @Service
 public class FuncionarioService {
-    @Autowired
-    private FuncionarioRepository funcionarioRepository;
+
+    private final FuncionarioRepository funcionarioRepository;
+
+    public FuncionarioService(FuncionarioRepository funcionarioRepository) {
+        this.funcionarioRepository = funcionarioRepository;
+    }
 
     public Iterable<Funcionario> findAll() {
         return funcionarioRepository.findAll();
@@ -24,16 +28,25 @@ public class FuncionarioService {
     //read
     public Funcionario readFuncionarioID(int id) {
         Optional<Funcionario> optionalFuncionario = funcionarioRepository.findById(id);
-        return optionalFuncionario.orElseThrow(() -> new RuntimeException("ID não encontrado"));
+        return optionalFuncionario.orElseThrow(() -> new NotFoundException("ID não encontrado"));
     }
 
     //update
-    public void updateFuncionario(Funcionario funcionario) {
-        funcionarioRepository.save(funcionario);
+    public Funcionario updateFuncionario(int id, Funcionario funcionario) {
+        Optional<Funcionario> optionalFuncionario = funcionarioRepository.findById(id);
+        if (!optionalFuncionario.isPresent()) {
+            throw new NotFoundException("ID não encontrado");
+        }
+        return funcionarioRepository.save(funcionario);
     }
 
     //delete
     public void deleteFuncionarioID(int id) {
+        Optional<Funcionario> optionalFuncionario = funcionarioRepository.findById(id);
+        if (!optionalFuncionario.isPresent()) {
+            throw new NotFoundException("ID não encontrado");
+        }
         funcionarioRepository.deleteById(id);
     }
+
 }
